# perf: only apply lazy cjs module transform on cli and core